\name{causalMomFct-methods}
\docType{methods}

\alias{causalMomFct-methods}
\alias{causalMomFct}
\alias{causalMomFct,numeric,causalData-method}

\title{Methods for Function \code{causalMomFct} in Package \pkg{causalGel}}
\description{
 The methods computes the moment matrix.
}
\section{Methods}{
\describe{

\item{\code{signature(theta = "numeric", object = "causalData")}}{
}
}}
\keyword{methods}
\keyword{Balancing moments}
